OrionServer/requirements.txt
protobuf==3.9.1
grpcio==1.24.3
grpcio-tools==1.24.3
pyserial==3.4